fix: изменение контроллера и сервиса, добавление логов
