+++ src/main/java/school/faang/user_service/entity/UserProfilePic.java
@@ -1,9 +1,15 @@
 package school.faang.user_service.entity;
 
 import jakarta.persistence.Embeddable;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
 @Embeddable
 public class UserProfilePic {
     private String fileId;
